refactor(observability): 抽 experience.ts 反馈匹配器簇到 feedback-matchers.ts
把 experience.ts 末尾的「用户反馈文本匹配器」整块抽到独立模块:8 个公开函数
(find/has × 4 类信号)+ 12 个词表常量 + 3 个 helper(pushTermMatches /
isTextBoundary / isBenignNegativeFeedbackContext)+ TextMatchRange 类型,
共 -204 行;experience.ts 通过 re-export 维持对外签名兼容,renderer 走
feedback-projection facade 直接指向新模块。

阶段 3 PR1。词表与上下文规则零字节改动,Report 字段语义(negativeFeedbackCount
等四个 indicators)完全不变。
